Extract getSection and informationSetAsNotReceived to seperate file
These were also causing a dependency cycle when used in later commits
so we extract them to their own file.
