+++ server/utils/assessmentUtils.ts
@@ -2,7 +2,6 @@ import { HtmlItem, PageResponse, SummaryListItem, TableRow, Task, TextItem } fro
 import { format, differenceInDays, add } from 'date-fns'
 
 import { ApprovedPremisesAssessment as Assessment, ApprovedPremisesApplication } from '@approved-premises/api'
-import AssessApplication from '../form-pages/assess/assessApplication'
 import { tierBadge } from './personUtils'
 import { DateFormats } from './dateUtils'
 import { getArrivalDate, getResponseForPage } from './applicationUtils'
@@ -106,24 +105,6 @@ const requestedFurtherInformationTableRows = (assessments: Array<Assessment>): A
   })
 
   return rows
-}
-
-const informationSetAsNotReceived = (assessment: Assessment): boolean => {
-  if (assessment.status === 'pending' && assessment.data) {
-    const response = assessment.data?.['sufficient-information']?.['information-received']?.informationReceived
-    return response === 'no'
-  }
-  return false
-}
-
-const getSections = (assessment: Assessment) => {
-  let { sections } = Assess
-
-  if (informationSetAsNotReceived(assessment)) {
-    sections = sections.filter(section => section.name !== AssessApplication.name)
-  }
-
-  return sections
 }
 
 const assessmentLink = (assessment: Assessment): string => {
@@ -322,6 +303,4 @@ export {
   assessmentsApproachingDue,
   assessmentsApproachingDueBadge,
   formatDaysUntilDueWithWarning,
-  informationSetAsNotReceived,
-  getSections,
 }
